// back/src/modules/xContentCrawler/types.ts
export interface XUser {
    username: string;
    displayName: string | null;
    url: string | null;
    agentId?: string; // 添加 agentId 字段
  }
  
  /**
   * Extended user with posts
   */
  export interface Following extends XUser {
    posts: XPost[];
    postsCollected: number;
    lastUpdate: Date | null;
    error: string | null;
  }
  
  /**
   * X post structure
   */
  export interface XPost {
    id: string;
    agentId?: string; // 添加 agentId 字段
    url: string | null;
    timestamp: string | null;
    text: string | null;
    author: {
      username: string;
      displayName?: string | null;
    };
    metrics: {
      replies: number | null;
      retweets: number | null;
      likes: number | null;
      views: number | null;
    };
  }
// ...
  export interface ContentInsight {
    hasValue: boolean;
    category: 'trading_idea' | 'project_intro' | 'market_insight' | 'none';
    // keyDetails: string[];
    summary: string;
    source: string;
    username: string;
    agentId: string; // 添加 agentId 字段
    timestamp: string;
    entity: Entity;
    event: Event;
  }

  


  export interface XPostPayload {
  userId: string;
  agentId: string; // 添加 agentId 字段
  post_content: string;
  authorUsername: string;
  url?: string;
  timestamp?: string;
}


export interface Entity {
  name: string;
  context: string;
}
export interface Event {
  name: string;
  details: string;
}